+++ lib/config/webpack-hooks.ts
@@ -1,0 +1,6 @@
+import { getCustomConfigFile } from '../runtime/util/fs'
+import { WebpackHooks } from '../types'
+
+export const webpackHooks = getCustomConfigFile<WebpackHooks>('config/webpack-hooks', {})
+
+export default webpackHooks