+++ lib/config/webpack.client.debug.ts
@@ -1,6 +1,7 @@
 import * as webpack from 'webpack'
 import * as merge from 'webpack-merge'
 import clientDevConfig from './webpack.client.dev'
+import webpackHooks from './webpack-hooks'
 
 /** Webpack config for the client to improve debugging */
 const config: webpack.Configuration = merge(
@@ -8,6 +9,7 @@ const config: webpack.Configuration = merge(
     {
         devtool: 'source-map',
     },
+    webpackHooks.clientDebug || {},
 )
 
 export default config