+++ lib/config/webpack.server.debug.ts
@@ -1,6 +1,7 @@
 import * as webpack from 'webpack'
 import * as merge from 'webpack-merge'
 import serverDevConfig from './webpack.server.dev'
+import webpackHooks from './webpack-hooks'
 
 /** Webpack config for the server to enable debugging */
 const config: webpack.Configuration = merge(
@@ -12,6 +13,7 @@ const config: webpack.Configuration = merge(
             devtoolFallbackModuleFilenameTemplate: '[absolute-resource-path]?[hash]',
         },
     },
+    webpackHooks.serverDebug || {},
 )
 
 export default config